Implement the ChaCha block function
This implementation is optimized for SSE instructions, going out of its
way to operate on four-element rows instead of single elements. The inner
loop of the block function has been tweaked until the compiler was
convinced to generate SSE instructions. Unrolling the loop, or even having
a double round inside the loop, seems to confuse the compiler into moving
data into and out of SSE registers. (This is using the 1.6 compiler.)